requirements.txt
paho-mqtt==1.2
telepot==10.5
urllib3==1.20
wsgiref==0.1.2
